# cmd/kubectl-audit-server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Multi-cluster aggregation server for kubectl-audit: clusters push findings.json scans here, which are deduplicated and stored for cross-cluster triage, centralized knowledge base/exclusion rules, and automation rule evaluation.",
        "title": "kubectl-audit-server API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/triage": {
            "get": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "Returns every finding for one cluster merged with its triage state (status defaults to \"new\" when no triage entry exists yet).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "triage"
                ],
                "summary": "Get the merged findings+triage view",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cluster ID",
                        "name": "cluster_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Restrict to one ingest source, e.g. native or openreports",
                        "name": "source",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_server_http.triageViewResponse"
                        }
                    },
                    "400": {
                        "description": "missing or invalid cluster_id",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "missing or invalid admin token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "no such cluster",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/triage/bulk": {
            "post": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "Uploads a client's whole local triage.State for one source and one cluster in a single call. Every field in each entry is authoritative (full replace, not a partial merge). Entries whose fingerprint has no matching Finding are skipped and reported, not treated as a hard failure.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "triage"
                ],
                "summary": "Bulk-replace triage entries for one source",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cluster ID",
                        "name": "cluster_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Source and its full set of triage entries",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_server_http.bulkTriageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_server_http.bulkTriageResponse"
                        }
                    },
                    "400": {
                        "description": "invalid request body / missing source / missing or invalid cluster_id",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "missing or invalid admin token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "no such cluster",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/triage/{source}/{fingerprint}": {
            "patch": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "Partially updates the TriageEntry for one finding (source+fingerprint) of one cluster. Omitted fields keep their current value; a first update creates the entry with status \"new\" as the baseline.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "triage"
                ],
                "summary": "Update one triage entry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cluster ID",
                        "name": "cluster_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Ingest source, e.g. native or openreports",
                        "name": "source",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Finding fingerprint",
                        "name": "fingerprint",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_server_http.patchTriageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ivanhahanov_kubectl-audit_internal_storage.TriageEntry"
                        }
                    },
                    "400": {
                        "description": "invalid request body / missing or invalid cluster_id",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "missing or invalid admin token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "no such cluster / no such finding for this cluster/source/fingerprint",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_ivanhahanov_kubectl-audit_internal_storage.TriageEntry": {
            "type": "object",
            "properties": {
                "clusterID": {
                    "type": "string"
                },
                "fingerprint": {
                    "type": "string"
                },
                "jiraIssueKey": {
                    "type": "string"
                },
                "jiraIssueURL": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "reviewer": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_ivanhahanov_kubectl-audit_internal_storage.TriageStatus"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_ivanhahanov_kubectl-audit_internal_storage.TriageStatus": {
            "type": "string",
            "enum": [
                "new",
                "confirmed",
                "false_positive",
                "wont_fix",
                "duplicate",
                "needs_info",
                "resolved"
            ],
            "x-enum-varnames": [
                "TriageStatusNew",
                "TriageStatusConfirmed",
                "TriageStatusFalsePositive",
                "TriageStatusWontFix",
                "TriageStatusDuplicate",
                "TriageStatusNeedsInfo",
                "TriageStatusResolved"
            ]
        },
...
        "internal_server_http.bulkTriageEntry": {
            "type": "object",
            "properties": {
                "fingerprint": {
                    "type": "string"
                },
                "jiraIssueKey": {
                    "type": "string"
                },
                "jiraIssueUrl": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "reviewer": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "internal_server_http.bulkTriageRequest": {
            "type": "object",
            "properties": {
                "entries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_server_http.bulkTriageEntry"
                    }
                },
                "source": {
                    "type": "string"
                }
            }
        },
        "internal_server_http.bulkTriageResponse": {
            "type": "object",
            "properties": {
                "skipped": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_server_http.bulkTriageSkip"
                    }
                },
                "updated": {
                    "type": "integer"
                }
            }
        },
        "internal_server_http.bulkTriageSkip": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "fingerprint": {
                    "type": "string"
                }
            }
        },
...
        "internal_server_http.patchTriageRequest": {
            "type": "object",
            "properties": {
                "jiraIssueKey": {
                    "type": "string"
                },
                "jiraIssueUrl": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "reviewer": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "internal_server_http.triageResourceRef": {
            "type": "object",
            "properties": {
                "apiVersion": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                }
            }
        },
        "internal_server_http.triageView": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "dedupKey": {
                    "description": "DedupKey mirrors findings.Finding.DedupKey — see that field's doc\ncomment. Without it, the TUI's bulk-triage collapsing has to bucket\npurely on Message text, which silently lumps together findings whose\nMessage happens to be identical but whose Resource genuinely differs\n(e.g. a check with a fixed, non-resource-specific Message template).",
                    "type": "string"
                },
                "fingerprint": {
                    "type": "string"
                },
                "firstSeen": {
                    "type": "string"
                },
                "jiraIssueKey": {
                    "type": "string"
                },
                "jiraIssueUrl": {
                    "type": "string"
                },
                "lastSeen": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "policyId": {
                    "type": "string"
                },
                "resource": {
                    "$ref": "#/definitions/internal_server_http.triageResourceRef"
                },
                "reviewer": {
                    "type": "string"
                },
                "severity": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "internal_server_http.triageViewResponse": {
            "type": "object",
            "properties": {
                "entries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_server_http.triageView"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "AdminAuth": {
            "description": "Admin bearer token (env ADMIN_TOKEN). Send as \"Bearer \u003ctoken\u003e\". Required for cluster registration and every organization-level config endpoint (knowledge base, exclusion rules, automation rules, triage).",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "ClusterAuth": {
            "description": "Per-cluster bearer token minted by POST /clusters. Send as \"Bearer \u003ctoken\u003e\". Required for ingest endpoints only — scoped to that cluster's own data, and does not grant triage read/write (see AdminAuth).",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}